Restructured page for displaying the share link
- make sure it can take multiple params with the first param being experience id.
- Restructured to code to make it more modular
